Move candletool headers into include
