+++ candletool/include/candletool/configHelpers.hpp
