+++ candletool/include/candletool/md_cli.hpp
@@ -4,7 +4,7 @@
 #include "candle/types/mab_types.hpp"
 #include "logger.hpp"
 #include "candle/MD/MD.hpp"
-#include "utilities.hpp"
+#include "candletool/utilities.hpp"
 #include <algorithm>
 #include <memory>
 #include <optional>
