+++ candletool/include/candletool/mdco_config_adapter.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "logger.hpp"
-#include "md_cfg_map.hpp"
+#include "candletool/md_cfg_map.hpp"
 #include "candle/objectDictionary/edsEntry.hpp"
 #include "candle/MD/MDCO.hpp"
