+++ candletool/include/candletool/web_file_module/curl_handler.hpp
@@ -5,7 +5,7 @@
 
 #include "candle/types/mab_types.hpp"
 #include "logger.hpp"
-#include "web_file.hpp"
+#include "candletool/web_file_module/web_file.hpp"
 
 namespace mab
 {
