+++ candletool/src/canLoader.cpp
@@ -1,8 +1,9 @@
-#include "canLoader.hpp"
+#include "candletool/canLoader.hpp"
+
 #include "candle/can_bootloader/can_bootloader.hpp"
 #include "candle/communication_device/candle.hpp"
-#include "mabFileParser.hpp"
-#include "mab_crc.hpp"
+#include "candletool/mabFileParser.hpp"
+#include "candletool/mab_crc.hpp"
 
 #include <array>
 #include <cmath>
