+++ candletool/src/mabFileParser.cpp
@@ -1,6 +1,8 @@
-#include "mabFileParser.hpp"
-#include "mini/ini.h"
+#include "candletool/mabFileParser.hpp"
+
 #include <cstring>
+
+#include "mini/ini.h"
 
 bool                          hexStringToBytes(u8 buffer[], u32 bufferLen, const std::string& str);
 MabFileParser::TargetDevice_E parseTargetDevice(std::string tag);
