+++ candletool/src/main.cpp
@@ -1,18 +1,18 @@
 #include <cstdlib>
 #include <memory>
 #include <string>
+
 #include "candle/communication_device/candle.hpp"
-#include "candle_cli.hpp"
+#include "candletool/candle_cli.hpp"
 #include "candle/communication_device/candle_types.hpp"
-#include "configHelpers.hpp"
+#include "candletool/configHelpers.hpp"
 #include "logger.hpp"
 #include "candle/types/mab_types.hpp"
-#include "md_cli.hpp"
+#include "candletool/md_cli.hpp"
 #include "CLI/CLI.hpp"
-#include "pds_cli.hpp"
-#include "mdco_cli.hpp"
-
-#include "utilities.hpp"
+#include "candletool/pds_cli.hpp"
+#include "candletool/mdco_cli.hpp"
+#include "candletool/utilities.hpp"
 
 //     ___     _     _  _      _   _         _____               _
 //    / __|   /_\   | \| |  __| | | |  ___  |_   _|  ___   ___  | |
