+++ candletool/src/mdco_cli.cpp
@@ -1,4 +1,5 @@
-#include "mdco_cli.hpp"
+#include "candletool/mdco_cli.hpp"
+
 #include <fcntl.h>
 #include <array>
 #include <chrono>
@@ -13,6 +14,7 @@
 #include <string_view>
 #include <thread>
 #include <vector>
+
 #include "CLI/CLI.hpp"
 #include "candle/MD/MDCO.hpp"
 #include "candle/communication_device/candle.hpp"
@@ -20,9 +22,9 @@
 #include "candle/objectDictionary/edsEntry.hpp"
 #include "candle/objectDictionary/edsParser.hpp"
 #include "candle/types/mab_types.hpp"
-#include "md_cfg_map.hpp"
+#include "candletool/md_cfg_map.hpp"
 #include "mini/ini.h"
-#include "mdco_config_adapter.hpp"
+#include "candletool/mdco_config_adapter.hpp"
 
 using namespace mab;
