+++ candletool/src/mdco_config_adapter.cpp
@@ -1,4 +1,4 @@
-#include "mdco_config_adapter.hpp"
+#include "candletool/mdco_config_adapter.hpp"
 
 namespace mab
 {
