+++ candletool/src/utilities.cpp
@@ -1,4 +1,5 @@
-#include "utilities.hpp"
+#include "candletool/utilities.hpp"
+
 #include <algorithm>
 #include <cctype>
 #include <variant>
